+++ app/src/main/java/org/fossasia/susi/ai/settings/ISettingsPresenter.kt
@@ -1,14 +1,14 @@
 package org.fossasia.susi.ai.settings
 
 import android.content.Context
-import android.content.SharedPreferences
 
 /**
  * Created by mayanktripathi on 07/07/17.
  */
+
 interface ISettingsPresenter {
 
-    fun onAttach(settingView: ISettingsView, preferences: SharedPreferences)
+    fun onAttach(chatSettingsFragment: ChatSettingsFragment)
 
     fun deleteMsg()
 
@@ -17,5 +17,7 @@ interface ISettingsPresenter {
     fun setTheme(string: String)
 
     fun enableMic(context: Context): Boolean
+
+    fun onDetach()
 
 }
