+++ app/src/main/java/org/fossasia/susi/ai/settings/ISettingsView.kt
@@ -5,11 +5,9 @@ import android.support.v4.app.FragmentActivity
 /**
  * Created by mayanktripathi on 07/07/17.
  */
+
 interface ISettingsView {
 
     fun showAlert(activity: FragmentActivity)
 
-    fun enableMic(): Boolean?
-
-    fun setTheme(string: String)
 }
