+++ app/src/main/java/org/fossasia/susi/ai/settings/SettingsPresenter.kt
@@ -1,7 +1,6 @@
 package org.fossasia.susi.ai.settings
 
 import android.content.Context
-import android.content.SharedPreferences
 
 /**
  * Created by mayanktripathi on 07/07/17.
@@ -11,12 +10,10 @@ class SettingsPresenter: ISettingsPresenter{
 
     var settingView: ISettingsView? = null
     var settingInteractor: SettingsInteractor? = null
-    var prefs: SharedPreferences? = null
 
-    override fun onAttach(settingView: ISettingsView, preferences: SharedPreferences) {
+    override fun onAttach(chatSettingsFragment: ChatSettingsFragment) {
         this.settingView = settingView
         this.settingInteractor = SettingsInteractor()
-        this.prefs = preferences
     }
 
     override fun deleteMsg() {
@@ -24,15 +21,19 @@ class SettingsPresenter: ISettingsPresenter{
     }
 
     override fun getThemes(): String? {
-        return settingInteractor?.getTheme(prefs!!)
+        return settingInteractor?.getTheme()
     }
 
     override fun setTheme(string: String) {
-        settingInteractor?.setTheme(string, prefs!!)
+        settingInteractor?.setTheme(string)
     }
 
     override fun enableMic(context: Context): Boolean {
         return settingInteractor?.setEnableMic(context)!!
+    }
+
+    override fun onDetach() {
+        settingView = null
     }
 
 }
