+++ app/src/main/res/layout/activity_settings.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/chat_pref"
-    android:name="org.fossasia.susi.ai.settings.SettingsActivity$ChatSettingsFragment"
+    android:name="org.fossasia.susi.ai.settings.ChatSettingsFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="org.fossasia.susi.ai.settings.SettingsActivity"/>
